requirements.txt
docker
dnspython
environs